moved the lineage_db_update to a util function
