+++ crates/router/src/core/user.rs
@@ -3185,29 +3185,11 @@ pub async fn switch_org_for_user(
             .clone(),
     };
 
-    tokio::spawn({
-        let state = state.clone();
-        let lineage_context = lineage_context.clone();
-        let user_id = user_from_token.user_id.clone();
-        async move {
-            if let Err(e) = state
-                .global_store
-                .update_user_by_user_id(
-                    &user_id,
-                    diesel_models::user::UserUpdate::LineageContextUpdate { lineage_context },
-                )
-                .await
-            {
-                logger::error!(
-                    "Failed to update lineage context for user {}: {:?}",
-                    user_id,
-                    e
-                );
-            } else {
-                logger::debug!("Successfully updated lineage context for user {}", user_id);
-            }
-        }
-    });
+    utils::user::spawn_async_lineage_context_update_to_db(
+        &state,
+        &user_from_token.user_id,
+        lineage_context,
+    );
 
     let token = utils::user::generate_jwt_auth_token_with_attributes(
         &state,
@@ -3417,27 +3399,11 @@ pub async fn switch_merchant_for_user_in_org(
             .clone(),
     };
 
-    tokio::spawn({
-        let state = state.clone();
-        let lineage_context = lineage_context.clone();
-        let user_id = user_from_token.user_id.clone();
-        async move {
-            let _ = state
-                .global_store
-                .update_user_by_user_id(
-                    &user_id,
-                    diesel_models::user::UserUpdate::LineageContextUpdate { lineage_context },
-                )
-                .await
-                .map_err(|e| {
-                    logger::error!(
-                        "Failed to update lineage context for user {}: {:?}",
-                        user_id,
-                        e
-                    );
-                });
-        }
-    });
+    utils::user::spawn_async_lineage_context_update_to_db(
+        &state,
+        &user_from_token.user_id,
+        lineage_context,
+    );
 
     let token = utils::user::generate_jwt_auth_token_with_attributes(
         &state,
@@ -3568,27 +3534,11 @@ pub async fn switch_profile_for_user_in_org_and_merchant(
             .clone(),
     };
 
-    tokio::spawn({
-        let state = state.clone();
-        let lineage_context = lineage_context.clone();
-        let user_id = user_from_token.user_id.clone();
-        async move {
-            let _ = state
-                .global_store
-                .update_user_by_user_id(
-                    &user_id,
-                    diesel_models::user::UserUpdate::LineageContextUpdate { lineage_context },
-                )
-                .await
-                .map_err(|e| {
-                    logger::error!(
-                        "Failed to update lineage context for user {}: {:?}",
-                        user_id,
-                        e
-                    );
-                });
-        }
-    });
+    utils::user::spawn_async_lineage_context_update_to_db(
+        &state,
+        &user_from_token.user_id,
+        lineage_context,
+    );
 
     let token = utils::user::generate_jwt_auth_token_with_attributes(
         &state,
